Move the edit function in the edit file
